+++ core/src/server/grpc_impl/GrpcRequestTask.cpp
@@ -29,7 +29,7 @@
 #include "scheduler/SchedInst.h"
 #include "server/DBWrapper.h"
 #include "server/Server.h"
-#include "src/config.h"
+#include "src/version.h"
 #include "utils/CommonUtil.h"
 #include "utils/Log.h"
 #include "utils/TimeRecorder.h"